+++ app/src/main/java/ash/app/journal/ui/data/JournalDatabase.kt
@@ -5,7 +5,11 @@ import androidx.room.RoomDatabase
 import ash.app.journal.ui.models.JournalEntry
 import ash.app.journal.ui.models.LinkMetadataEntity
 
-@Database(entities = [JournalEntry::class, LinkMetadataEntity::class], version = 3, exportSchema = false)
+@Database(
+    entities = [JournalEntry::class, LinkMetadataEntity::class],
+    version = 4,
+    exportSchema = false
+)
 abstract class JournalDatabase : RoomDatabase() {
     abstract fun journalDao(): JournalDao
     abstract fun linkMetadataDao(): LinkMetadataDao